+++ src/icons/arrow/ArrowNewLineIcon.vue
@@ -1,0 +1,11 @@
+<template>
+  <svg fill="none" height="20" viewBox="0 0 20 20" width="20" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M6.99219 0.992186L6.99219 8C6.99219 9.65685 8.33533 11 9.99219 11L17 11"
+      stroke="currentColor"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      stroke-width="1.5"
+    />
+  </svg>
+</template>